requirements.txt
asgiref==3.3.1
Django==3.1.3
django-cors-headers==3.6.0
djangorestframework==3.12.2
boto3==1.14.5
botocore==1.17.5
dj-database-url==0.5.0
django-filter==2.3.0
django-heroku==0.3.1
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0